100baseT1: fixed push constant sizing issue causing one of the scrambler polynomials to not work

+++ scopeprotocols/Ethernet100BaseT1Decoder.h
@@ -51,7 +51,7 @@ public:
 	uint32_t	len;
 	uint32_t	samplesPerThread;
 	uint32_t	maxOutputPerThread;
-	uint8_t		masterMode;
+	uint32_t	masterMode;
 };
 
 class BaseT1DecodeConstants
